Move RepairTextFunction to ./repair-text.ts
