+++ packages/ai/src/generate-object/generate-object.ts
@@ -42,19 +42,9 @@ import { LanguageModelUsage } from '../types/usage';
 import { GenerateObjectResult } from './generate-object-result';
 import { getOutputStrategy } from './output-strategy';
 import { validateObjectGenerationInput } from './validate-object-generation-input';
+import { RepairTextFunction } from './repair-text';
 
 const originalGenerateId = createIdGenerator({ prefix: 'aiobj', size: 24 });
-
-/**
-A function that attempts to repair the raw output of the mode
-to enable JSON parsing.
-
-Should return the repaired text or null if the text cannot be repaired.
-     */
-export type RepairTextFunction = (options: {
-  text: string;
-  error: JSONParseError | TypeValidationError;
-}) => Promise<string | null>;
 
 /**
 Generate a structured, typed object for a given prompt and schema using a language model.
