+++ packages/ai/src/generate-object/index.ts
@@ -1,5 +1,5 @@
 export { generateObject } from './generate-object';
-export type { RepairTextFunction } from './generate-object';
+export type { RepairTextFunction } from './repair-text';
 export type { GenerateObjectResult } from './generate-object-result';
 export { streamObject } from './stream-object';
 export type { StreamObjectOnFinishCallback } from './stream-object';
